+++ Tests/scripts/hook_validations/script.py
@@ -1,12 +1,5 @@
-import os
-import yaml
-import requests
-
-from Tests.scripts.constants import CONTENT_GITHUB_LINK, PYTHON_SUBTYPES
-from Tests.test_utils import print_error, print_warning, get_yaml, server_version_compare
-
-# disable insecure warnings
-requests.packages.urllib3.disable_warnings()
+from Tests.scripts.constants import PYTHON_SUBTYPES
+from Tests.test_utils import print_error, get_yaml, get_remote_file, server_version_compare, get_dockerimage45
 
 
 class ScriptValidator(object):
@@ -26,19 +19,9 @@ class ScriptValidator(object):
 
         if check_git:
             self.current_script = get_yaml(file_path)
-            # The replace in the end is for Windows support
-            if old_file_path:
-                git_hub_path = os.path.join(CONTENT_GITHUB_LINK, old_git_branch, old_file_path).replace("\\", "/")
-            else:
-                git_hub_path = os.path.join(CONTENT_GITHUB_LINK, old_git_branch, file_path).replace("\\", "/")
 
-            try:
-                res = requests.get(git_hub_path, verify=False)
-                res.raise_for_status()
-                self.old_script = yaml.safe_load(res.content)
-            except Exception as e:
-                print_warning("{}\nCould not find the old script please make sure that you did not break "
-                              "backward compatibility".format(str(e)))
+            old_script_file = old_file_path or file_path
+            self.old_script = get_remote_file(old_script_file, old_git_branch)
 
     @classmethod
     def _is_sub_set(cls, supposed_bigger_list, supposed_smaller_list):
@@ -178,9 +161,11 @@ class ScriptValidator(object):
         """Check if the docker image as been changed."""
         # Unnecessary to check docker image only on 5.0 and up
         if server_version_compare(self.old_script.get('fromversion', '0'), '5.0.0') < 0:
-            if self.old_script.get('dockerimage', "") != self.current_script.get('dockerimage', ""):
+            old_docker = get_dockerimage45(self.old_script)
+            new_docker = get_dockerimage45(self.current_script)
+            if old_docker != new_docker:
                 print_error("Possible backwards compatibility break, You've changed the docker for the file {}"
-                            " this is not allowed.".format(self.file_path))
+                            " this is not allowed. Old: {}. New: {}".format(self.file_path, old_docker, new_docker))
                 return True
 
         return False